pyproject.toml
[tool.poetry]
name = "pyba"
version = "0.0.1"
description = "Automate online browsing using python and AI"
authors = ["pUrGe12 <achintya.jai@owasp.org>"]
keywords = [
    "browser-automations",
    "AI",
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
license = "MIT"
readme = "README.md"
packages = [{include = "pyba", from = "core"}]

[tool.poetry.dependencies]
python = ">=3.10"
pydantic = "^2.12.0"
google-genai = "^1.45.0"
openai = "^2.6.0"
bs4 = "^0.0.2"

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = ["pyba"]
no_inline_sort = true
skip = ["__init__.py"]

[tool.ruff]
line-length = 99
per-file-ignores = { "__init__.py" = ["ALL"]}

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
